requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.1
dj-database-url==1.2.0
Django==4.1.5
django-daraja==1.2.0
django-heroku==0.3.1
django-mpesa==2.0.14
django-phonenumber-field==7.0.2
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.4
mpesa==0.0.1
phonenumbers==8.13.8
phonenumberslite==8.13.8
Pillow==9.4.0
psycopg2==2.9.5
pycparser==2.21
python-decouple==3.7
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.15
whitenoise==6.3.0
